feat: add default error with status support
